struct(T4): extract MusicGenre into game/genre.rs
Pure move so genre is a single owned module ahead of the world split
and future ability weights. Canonical path is crate::game::genre::MusicGenre
with no world re-export; random/random_trending stay crate-visible for
scene generation. Marks T4 done on the structure handoff board.
