+++ src/game/world.rs
@@ -1,5 +1,6 @@
 use crate::data_loader::{GameDataFiles, RecordLabel};
 use crate::game::band::Band; // Added import for Band
+use crate::game::genre::MusicGenre;
 use crate::game::timeline::MusicTimeline;
 use rand::Rng;
 use serde::{Deserialize, Serialize};
@@ -93,79 +94,6 @@ pub enum MusicTrend {
     Punk,
     Alternative,
     Electronic,
-}
-
-// Ensure MusicGenre can be used as a HashMap key
-#[derive(Debug, Clone, Default, Serialize, Deserialize, PartialEq, Eq, Hash)]
-pub enum MusicGenre {
-    #[default]
-    Rock,
-    Pop,
-    Metal,
-    Punk,
-    Alternative,
-    Electronic,
-    Folk,
-    Jazz,
-}
-
-impl MusicGenre {
-    pub const ALL: [MusicGenre; 8] = [
-        MusicGenre::Rock,
-        MusicGenre::Pop,
-        MusicGenre::Metal,
-        MusicGenre::Punk,
-        MusicGenre::Alternative,
-        MusicGenre::Electronic,
-        MusicGenre::Folk,
-        MusicGenre::Jazz,
-    ];
-
-    pub fn name(&self) -> &'static str {
-        match self {
-            MusicGenre::Rock => "Rock",
-            MusicGenre::Pop => "Pop",
-            MusicGenre::Metal => "Metal",
-            MusicGenre::Punk => "Punk",
-            MusicGenre::Alternative => "Alternative",
-            MusicGenre::Electronic => "Electronic",
-            MusicGenre::Folk => "Folk",
-            MusicGenre::Jazz => "Jazz",
-        }
-    }
-
-    /// Keys this genre answers to in markets.json's genre_era_modifiers.
-    pub fn aliases(&self) -> &'static [&'static str] {
-        match self {
-            MusicGenre::Rock => &["rock", "arena_rock"],
-            MusicGenre::Pop => &["pop", "synth_pop", "brit_pop", "disco"],
-            MusicGenre::Metal => &["metal", "hair_metal", "pop_metal"],
-            MusicGenre::Punk => &["punk", "post_punk"],
-            MusicGenre::Alternative => &["alternative", "grunge", "new_wave"],
-            MusicGenre::Electronic => &["synth_pop", "house", "new_wave", "disco"],
-            MusicGenre::Folk => &["folk"],
-            MusicGenre::Jazz => &["jazz", "blues"],
-        }
-    }
-
-    fn random(rng: &mut impl Rng) -> Self {
-        MusicGenre::ALL[rng.gen_range(0..MusicGenre::ALL.len())].clone()
-    }
-
-    /// A genre that fits the current trends, if any does.
-    fn random_trending(timeline: &MusicTimeline, rng: &mut impl Rng) -> Self {
-        let trending = timeline.get_trending_genres();
-        let matching: Vec<MusicGenre> = MusicGenre::ALL
-            .iter()
-            .filter(|genre| trending.iter().any(|t| t.contains(genre.name())))
-            .cloned()
-            .collect();
-        if matching.is_empty() {
-            Self::random(rng)
-        } else {
-            matching[rng.gen_range(0..matching.len())].clone()
-        }
-    }
 }
 
 /// The scene never empties out or grows without bound.
@@ -806,12 +734,6 @@ impl std::fmt::Display for EconomicState {
             EconomicState::Growing => write!(f, "Growing"),
             EconomicState::Booming => write!(f, "Booming"),
         }
-    }
-}
-
-impl std::fmt::Display for MusicGenre {
-    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
-        write!(f, "{}", self.name())
     }
 }
